[jest-jasmine2] Split `jasmine-light` over multiple files. (#3159)
* [jest-jasmine2] Split `jasmine-light` over multiple files.

* Don't wrap things that don't need to be wrapped.

* Don't wrap things that don't need to be wrapped (again).

* Lint.

* Move source to `jasmine`, alter license.
