+++ packages/jest-jasmine2/src/jasmine/buildExpectationResult.js
@@ -1,0 +1,84 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @flow
+ */
+// This file is a heavily modified fork of Jasmine. The original license of the code:
+/*
+Copyright (c) 2008-2016 Pivotal Labs
+
+Permission is hereby granted, free of charge, to any person obtaining
+a copy of this software and associated documentation files (the
+"Software"), to deal in the Software without restriction, including
+without limitation the rights to use, copy, modify, merge, publish,
+distribute, sublicense, and/or sell copies of the Software, and to
+permit persons to whom the Software is furnished to do so, subject to
+the following conditions:
+
+The above copyright notice and this permission notice shall be
+included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+/* eslint-disable sort-keys */
+'use strict';
+
+function buildExpectationResult(options) {
+  const messageFormatter = options.messageFormatter || function() {};
+  const stackFormatter = options.stackFormatter || function() {};
+
+  const result = {
+    matcherName: options.matcherName,
+    message: message(),
+    stack: stack(),
+    passed: options.passed,
+    // CUSTOM JEST CHANGE: we pass error message to the result.
+    error: options.error,
+  };
+
+  if (!result.passed) {
+    result.expected = options.expected;
+    result.actual = options.actual;
+  }
+
+  return result;
+
+  function message() {
+    if (options.passed) {
+      return 'Passed.';
+    } else if (options.message) {
+      return options.message;
+    } else if (options.error) {
+      return messageFormatter(options.error);
+    }
+    return '';
+  }
+
+  function stack() {
+    if (options.passed) {
+      return '';
+    }
+
+    let error = options.error;
+    if (!error) {
+      try {
+        throw new Error(message());
+      } catch (e) {
+        error = e;
+      }
+    }
+    return stackFormatter(error);
+  }
+}
+
+module.exports = buildExpectationResult;
